Issue #SB-27408 | Assessment archival Base and sub class changes
