+++ data-products/src/main/scala/org/sunbird/analytics/archival/BaseArchivalJob.scala
@@ -31,8 +31,6 @@ case class ArchivalMetrics(batchId: Option[String],
 
 trait BaseArchivalJob extends BaseReportsJob with IJob with ArchivalMetaDataStoreJob with Serializable {
 
-  private val partitionCols = List("batch_id", "year", "week_of_year")
-  private val columnWithOrder = List("course_id", "batch_id", "user_id", "content_id", "attempt_id", "created_on", "grand_total", "last_attempted_on", "total_max_score", "total_score", "updated_on", "question")
   val cassandraUrl = "org.apache.spark.sql.cassandra"
 
   def main(config: String)(implicit sc: Option[SparkContext] = None, fc: Option[FrameworkContext] = None): Unit = {
@@ -64,122 +62,33 @@ trait BaseArchivalJob extends BaseReportsJob with IJob with ArchivalMetaDataStor
   }
 
 //  def dataFilter(): Unit = {}
-//  def dateFormat(): String;
+
   def getClassName: String;
 
   def execute()(implicit spark: SparkSession, fc: FrameworkContext, config: JobConfig): Unit = {
     val modelParams = config.modelParams.getOrElse(Map[String, Option[AnyRef]]());
     val requestConfig = JSONUtils.deserialize[Request](JSONUtils.serialize(modelParams.getOrElse("request", Request).asInstanceOf[Map[String,AnyRef]]))
-    val archivalTable = requestConfig.archivalTable
-    val archivalKeyspace = requestConfig.keyspace.getOrElse(AppConf.getConfig("sunbird.courses.keyspace"))
-
-    val batchId: String = requestConfig.batchId.getOrElse("")
-    val date: String  = requestConfig.date.getOrElse("")
     val mode: String = modelParams.getOrElse("mode","archive").asInstanceOf[String]
-
-    println("modelParams: " + modelParams)
-    println("archival request: " + requestConfig)
-    val archivalTableData: DataFrame = getArchivalData(archivalTable, archivalKeyspace,Option(batchId),Option(date))
-    println("archivalTableData ")
-    archivalTableData.show(false)
 
     mode.toLowerCase() match {
       case "archival" =>
-        archiveData(archivalTableData, requestConfig)
+        archiveData(requestConfig)
       case "delete" =>
-        deleteArchivedData(archivalTableData,requestConfig)
+        deleteArchivedData(requestConfig)
     }
   }
 
-  def archiveData(data: DataFrame, requestConfig: Request)(implicit spark: SparkSession, fc: FrameworkContext, config: JobConfig): Unit = {
-    val requests = getRequests(jobId, requestConfig.batchId)
-    println("requestLength: " + requests.length)
-    try {
-      var dataDF = processArchival(data, requestConfig)
-      if(requests.length > 0) {
-        for (request <- requests) {
-          if (request.archival_status.equals("SUCCESS")) {
-            val request_data = JSONUtils.deserialize[Map[String, AnyRef]](request.request_data)
-            dataDF = dataDF.filter(
-              col("batch_id").equalTo(request.batch_id) &&
-              concat(col("year"), lit("-"), col("week_of_year")) =!= lit(request_data.get("year").get + "-" + request_data.get("week").get)
-            )
-          }
-        }
-      }
-
-      val archiveBatchList = dataDF.groupBy(partitionCols.head, partitionCols.tail: _*).count().collect()
-
-      val batchesToArchive: Map[String, Array[BatchPartition]] = archiveBatchList.map(f => BatchPartition(f.get(0).asInstanceOf[String], Period(f.get(1).asInstanceOf[Int], f.get(2).asInstanceOf[Int]))).groupBy(_.batchId)
-
-      archiveBatches(batchesToArchive, dataDF, requestConfig)
-    } catch {
-      case ex: Exception =>
-        ex.printStackTrace()
-    }
+  def arhivalDateFormat(combinationList: List[String]): String = {
+    combinationList.mkString("-")
   }
 
-  def generatePeriodInData(data: DataFrame): DataFrame = {
-    data.withColumn("updated_on", to_timestamp(col("updated_on")))
-      .withColumn("year", year(col("updated_on")))
-      .withColumn("week_of_year", weekofyear(col("updated_on")))
-      .withColumn("question", to_json(col("question")))
-  }
+  def archiveData(requestConfig: Request)(implicit spark: SparkSession, fc: FrameworkContext, config: JobConfig): Unit
 
-  def archiveBatches(batchesToArchive: Map[String, Array[BatchPartition]], data: DataFrame, requestConfig: Request)(implicit config: JobConfig): Unit = {
-    batchesToArchive.foreach(batches => {
-      val processingBatch = new AtomicInteger(batches._2.length)
-      JobLogger.log(s"Started Processing to archive the data", Some(Map("batch_id" -> batches._1, "total_part_files_to_archive" -> processingBatch)))
+  def archiveBatches(batchesToArchive: Map[String, Array[BatchPartition]], data: DataFrame, requestConfig: Request)(implicit config: JobConfig): Unit
 
-      // Loop through the week_num & year batch partition
-      batches._2.map((batch: BatchPartition) => {
-        val filteredDF = data.filter(col("batch_id") === batch.batchId && col("year") === batch.period.year && col("week_of_year") === batch.period.weekOfYear).select(columnWithOrder.head, columnWithOrder.tail: _*)
-        val collectionId = filteredDF.first().getAs[String]("course_id")
-        var archivalRequest = getRequest(collectionId, batch.batchId, batch.period.year, batch.period.weekOfYear)
-
-        if (archivalRequest == null) {
-          val request_data = JSONUtils.deserialize[Map[String, AnyRef]](JSONUtils.serialize(requestConfig)) ++ Map[String, Int](
-            "week" -> batch.period.weekOfYear,
-            "year"-> batch.period.year
-          )
-          archivalRequest = ArchivalRequest("", batch.batchId, collectionId, Option(getReportKey), jobId, None, None, null, null, None, Option(0), JSONUtils.serialize(request_data), None)
-        }
-
-        try {
-          val urls = upload(filteredDF, batch) // Upload the archived files into blob store
-          archivalRequest.blob_url = Option(urls)
-          JobLogger.log(s"Data is archived and Processing the remaining part files ", None, Level.INFO)
-          markRequestAsSuccess(archivalRequest, requestConfig)
-        } catch {
-          case ex: Exception => {
-            markArchivalRequestAsFailed(archivalRequest, ex.getLocalizedMessage)
-          }
-        }
-      }).foreach((archivalRequest: ArchivalRequest) => {
-        upsertRequest(archivalRequest)
-      })
-
-      JobLogger.log(s"${batches._1} is successfully archived", Some(Map("batch_id" -> batches._1)), Level.INFO)
-    })
-  }
-
-  def deleteArchivedData(data: DataFrame, archivalRequest: Request): Unit = {
-
-  }
+  def deleteArchivedData(archivalRequest: Request): Unit
 
   def processArchival(archivalTableData: DataFrame, archiveRequest: Request)(implicit spark: SparkSession, fc: FrameworkContext, config: JobConfig): DataFrame;
-
-  def getArchivalData(table: String, keyspace: String, batchId: Option[String], date: Option[String])(implicit spark: SparkSession, fc: FrameworkContext): DataFrame = {
-    val archivalTableSettings = Map("table" -> table, "keyspace" -> keyspace, "cluster" -> "LMSCluster")
-    val archivalDBDF = loadData(archivalTableSettings, cassandraUrl, new StructType())
-    val batchIdentifier = batchId.getOrElse(null)
-
-    if (batchIdentifier.nonEmpty) {
-      archivalDBDF.filter(col("batch_id") === batchIdentifier).persist()
-    } else {
-      archivalDBDF
-    }
-  }
 
   def getWeekAndYearVal(date: String): Period = {
     if (null != date && date.nonEmpty) {
